% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoBRT-package.R
\docType{package}
\name{EcoBRT-package}
\alias{EcoBRT}
\alias{EcoBRT-package}
\title{EcoBRT: What the Package Does (Title Case)}
\description{
Collection of functions to generate Boosted Regression Trees using Gradient Boosting techniques for ecological data in R
}
\keyword{internal}
